textme_env/requirements.txt
Django~=1.11
djangorestframework~=3.8
djangorestframework-jwt~=1.11